api/requirements.txt
fastapi
uvicorn
python-multipart
numpy
pandas
scikit-learn
joblib
tensorflow
Pillow
google-cloud-storage
